✨ Move "Finish Voting" to voting screen
