+++ src/components/Accordion/Accordion.tsx
@@ -8,7 +8,7 @@ interface AccordionProps {
 }
 
 const Accordion: React.FC<AccordionProps> = ({ title, children }) => {
-	const [isOpen, setIsOpen] = useState(false)
+	const [isOpen, setIsOpen] = useState(true)
 
 	const toggleAccordion = () => {
 		setIsOpen(!isOpen)
