+++ src/components/list/list.css
@@ -3,4 +3,34 @@
   flex-direction: column;
   flex: 1;
   overflow: scroll;
+
+  .change-icons {
+    align-items: center;
+    justify-content: center;
+    padding-top: 20px;
+    gap: 20px;
+    display: flex;
+
+    img {
+      margin-right: 20px;
+      margin-left: 20px;
+      width: 35px;
+      height: 35px;
+      cursor: pointer;
+    }
+    button {
+      padding: 20px;
+      width: 100%;
+      height: 40px;
+      border-radius: 10px;
+      background-color: #1a73e8;
+      color: white;
+      border: none;
+      cursor: pointer;
+      display: flex;
+      align-items: center;
+      font-weight: bold;
+      font-size: 24px;
+    }
+  }
 }
